Split up API and Reducer functions
