+++ src/redux/configureStore.js
@@ -1,7 +1,7 @@
 import { createStore, applyMiddleware } from 'redux';
 import logger from 'redux-logger';
 import thunkMiddleware from 'redux-thunk';
-import covidDataReducer from './CovidAPI/covidAPI';
+import covidDataReducer from './CovidInfo/covidInfo';
 
 const store = createStore(
   covidDataReducer,
